+++ option.go
@@ -4,14 +4,16 @@ import "time"
 
 // Settings contains option to Conn.
 type Settings struct {
-	dialTimeout time.Duration
-	deadline    time.Duration
+	dialTimeout   time.Duration
+	deadline      time.Duration
+	maxCommandLen int
 }
 
 // DefaultSettings provides default deadline settings to Conn.
 var DefaultSettings = Settings{
-	dialTimeout: DefaultDialTimeout,
-	deadline:    DefaultDeadline,
+	dialTimeout:   DefaultDialTimeout,
+	deadline:      DefaultDeadline,
+	maxCommandLen: DefaultMaxCommandLen,
 }
 
 // Option allows to inject settings to Settings.
@@ -28,5 +30,12 @@ func SetDialTimeout(timeout time.Duration) Option {
 func SetDeadline(timeout time.Duration) Option {
 	return func(s *Settings) {
 		s.deadline = timeout
+	}
+}
+
+// SetMaxCommandLen injects maxCommandLen to Settings.
+func SetMaxCommandLen(maxCommandLen int) Option {
+	return func(s *Settings) {
+		s.maxCommandLen = maxCommandLen
 	}
 }